+++ e2e/wdio/core/fixtures/appData/pagination-contents.ts
@@ -1,72 +1,9 @@
-export const basicPaginationTestArr = [
-    'New page selected: 1',
-    'New page selected: 2',
-    'New page selected: 3',
-    'New page selected: 4',
-    'New page selected: 5'
-];
+export const basicPaginationTestArr = ['New page selected: 1', 'New page selected: 2', 'New page selected: 5'];
 export const itemPaginationTestArr = [
     'From 1 to 10. Total items 50',
     'From 11 to 20. Total items 50',
-    'From 21 to 30. Total items 50',
-    'From 31 to 40. Total items 50',
-    'From 41 to 50. Total items 50'
-];
-export const selectPaginationTestArr1 = [
-    'Showing 1-15 of 50',
-    'Showing 16-30 of 50',
-    'Showing 31-45 of 50',
-    'Showing 46-50 of 50'
-];
-export const selectPaginationTestArr2 = [
-    'Showing 1-4 of 50',
-    'Showing 5-8 of 50',
-    'Showing 49-50 of 50',
-    'Showing 49-50 of 50'
-];
-export const selectPaginationTestArr3 = [
-    'Showing 1-8 of 50',
-    'Showing 9-16 of 50',
-    'Showing 17-24 of 50',
-    'Showing 25-32 of 50',
-    'Showing 33-40 of 50',
-    'Showing 41-48 of 50',
-    'Showing 49-50 of 50'
-];
-export const selectPaginationTestArr4 = [
-    'Showing 1-16 of 50',
-    'Showing 17-32 of 50',
-    'Showing 33-48 of 50',
-    'Showing 49-50 of 50'
-];
-export const selectPaginationTestArr5 = [
-    'Showing 1-10 of 50',
-    'Showing 11-20 of 50',
-    'Showing 21-30 of 50',
-    'Showing 31-40 of 50',
-    'Showing 41-50 of 50'
-];
-export const selectPaginationTestArr6 = [
-    'Showing 1-2 of 50',
-    'Showing 3-4 of 50',
-    'Showing 49-50 of 50',
-    'Showing 49-50 of 50'
-];
-export const selectPaginationTestArr7 = [
-    'Showing 1-5 of 50',
-    'Showing 6-10 of 50',
-    'Showing 46-50 of 50',
-    'Showing 46-50 of 50'
-];
-export const selectPaginationTestText = 'Showing 1-50 of 50';
-export const playgroundPaginationItemArr = ['1', '2', '3', '4', '5'];
-export const playgroundPaginationItemArr2 = ['1', '2', '3', '4', '9'];
-export const playgroundPaginationItemArr3 = ['1', '6', '7', '8', '9'];
-export const playgroundPaginationTestText = [
-    'Showing 1-1 of 5',
-    'Showing 2-2 of 5',
-    'Showing 3-3 of 5',
-    'Showing 4-4 of 5',
-    'Showing 5-5 of 5'
+    'From 21 to 30. Total items 50'
 ];
 export const playgroundLabelArr = ['totalItems', 'itemsPerPage', 'currentPage', 'displayText'];
+export const resultsArr = ['4', '8', '16'];
+export const ResultsArr2 = ['25', '10', '1'];