+++ e2e/wdio/platform/tests/upload-collection.e2e-spec.ts
@@ -1,7 +1,6 @@
 import { UploadCollectionPo } from '../pages/upload-collection.po';
 
 import {
-    browserIsFirefox,
     click,
     getCurrentUrl,
     getElementArrayLength,
@@ -90,10 +89,6 @@ describe('Upload collection test suite', () => {
     });
 
     it('should check selected pages by clicking previous and next link for all examples', () => {
-        // skipped due to cannot reproduce failure, needs further investigation
-        if (getCurrentUrl().includes('localhost')) {
-            return;
-        }
         checkSelectedPagesByNextPrevious(defaultExample);
         checkSelectedPagesByNextPrevious(disableExample);
         checkSelectedPagesByNextPrevious(readonlyExample);
@@ -131,7 +126,8 @@ describe('Upload collection test suite', () => {
     });
 
     it('should check moving folders', () => {
-        if (browserIsFirefox()) {
+        // skipped due to cannot reproduce failure, needs further investigation
+        if (getCurrentUrl().includes('localhost')) {
             return;
         }
         checkMovingFolders(defaultExample);
@@ -146,8 +142,9 @@ describe('Upload collection test suite', () => {
         const movedFolderName = getText(selector + fileNameLabel);
         click(selector + tableItem);
         click(selector + checkbox, 1);
-        click(selector + ghostButton, 0);
-        pause(300);
+        click(selector + ghostButton);
+        pause(1000);
+
         const folderName = getText(listItemTitle, 1);
         click(listItem, 1);
         click(moveButton);